Refactored the project layout to make more sense
