+++ JavaFXProject/src/appElements/FxAlert.java
@@ -1,4 +1,4 @@
-package fxtools;
+package appElements;
 
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
