+++ JavaFXProject/src/appElements/ItemPopupControl.java
@@ -1,4 +1,4 @@
-package modalPopups;
+package appElements;
 
 import java.util.ArrayList;
