+++ JavaFXProject/src/application/StaticModalManager.java
@@ -2,6 +2,7 @@ package application;
 
 import java.io.IOException;
 
+import appElements.ItemPopupControl;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
@@ -9,7 +10,6 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
-import modalPopups.ItemPopupControl;
 
 public class StaticModalManager {
