+++ JavaFXProject/src/backend/ConnectFirebase.java
@@ -1,4 +1,4 @@
-package application;
+package backend;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
