+++ JavaFXProject/src/backend/Table.java
@@ -1,4 +1,4 @@
-package application;
+package backend;
 
 import java.util.HashMap;
 import java.util.Map.Entry;
@@ -9,74 +9,68 @@ import com.firebase.client.Firebase;
  *
  * @author Rich Cherngchaosil
  *
- * Object to store in the database
+ *         Object to store in the database
  *
  */
 public class Table {
 
 	/**
-	 * state:
-	 * 	0 - Ready
-	 * 	1 - Occ'd
-	 *  2 - Dirty
+	 * state: 0 - Ready 1 - Occ'd 2 - Dirty
 	 */
 	private int state;
 	private Firebase ref;
 	HashMap<String, Boolean> order;
 
-	public Table(){
+	public Table() {
 		state = 0;
 		order = new HashMap<String, Boolean>();
 
 	}
 
-
-
-	public int getTableState(){
+	public int getTableState() {
 		return state;
 	}
 
-
-	public void updateTableState(int s){
-		if(s < 0 || s > 2){
+	public void updateTableState(int s) {
+		if (s < 0 || s > 2) {
 			System.out.println("Unknown State, cannot update the state of the table");
 		}
 
 		state = s;
 	}
 
-	public void setRef(Firebase r){
+	public void setRef(Firebase r) {
 		ref = r;
 	}
 
-	public Firebase getRef(){
+	public Firebase getRef() {
 		return ref;
 	}
 
-	public void clearTable(){
+	public void clearTable() {
 		state = 0;
 	}
 
 	/** Ordering functions **/
 
-	public HashMap<String, Boolean> getOrder(){
+	public HashMap<String, Boolean> getOrder() {
 		return order;
 	}
 
-	public void addOrder(String food){
+	public void addOrder(String food) {
 		order.put(food, false);
 	}
 
-	public void served(String food){
+	public void served(String food) {
 		order.put(food, true);
 	}
 
-	public void updateOrder(HashMap<String, Boolean> newOrder){
+	public void updateOrder(HashMap<String, Boolean> newOrder) {
 		order = newOrder;
 	}
 
-	public void listFood(){
-		for(Entry<String, Boolean> entry: order.entrySet()){
+	public void listFood() {
+		for (Entry<String, Boolean> entry : order.entrySet()) {
 			System.out.println(entry.getKey());
 		}
 	}
